endpoint/requirements.txt
mysql-connector-python
Flask-JWT-Extended
flask-httpauth
